requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
python-dotenv==1.0.1
sqlalchemy==2.0.25
asyncpg==0.29.0
dashscope==1.14.0
jinja2==3.1.3
pandas